+++ lib/widgets/graph_component.dart
@@ -25,6 +25,7 @@ class GraphComponent extends FlameGame
   late GraphAlgorithm algorithm;
   late DataConvertor convertor;
   BuildContext context;
+  Options? options;
 
   ValueNotifier<double> scale = ValueNotifier(1);
   Vector2? pointLocation;
@@ -34,6 +35,7 @@ class GraphComponent extends FlameGame
     required this.algorithm,
     required this.context,
     required this.convertor,
+    this.options,
   });
 
   late Graph graph;